SberProject/search/requirements.txt
fastapi==0.78.0
pydantic==1.9.1
uvicorn[standard]==0.18.2
weaviate-client==3.5.1
